Moving REVIEW_STATUS enum to separated file and checking the status on the Review page to only display the submission for DRAFT reviews
