+++ src/utils/data/reviewStatus.ts
@@ -1,0 +1,7 @@
+export enum REVIEW_STATUS {
+  DRAFT = 'Draft',
+  PENDING = 'Pending',
+  CHANGES_REQUESTED = 'ChangesRequested',
+  SUBMITTED = 'Submitted',
+  LOCKED = 'Locked',
+}
