+++ src/utils/helpers/review/getReviewStartLabel.ts
@@ -1,12 +1,5 @@
 import strings from '../../constants'
-
-export enum REVIEW_STATUS {
-  DRAFT = 'Draft',
-  PENDING = 'Pending',
-  CHANGES_REQUESTED = 'ChangesRequested',
-  SUBMITTED = 'Submitted',
-  LOCKED = 'Locked',
-}
+import { REVIEW_STATUS } from '../../data/reviewStatus'
 
 const getReviewStartLabel = (status: string) => {
   // TODO: Not use strings here - use the type
